Remaking the trending words query
