+++ src/.eslintrc.json
@@ -59,7 +59,7 @@
     "networkGrowth_": "readonly",
     "exchangeFundsFlow_": "readonly",
     "tokenCirculation_": "readonly",
-    "trendingWords_": "readonly",
+    "emergingTrends_": "readonly",
     "fetchProjectFundamentals_": "readonly",
     "projectSocialData_": "readonly",
     "mvrvRatio_": "readonly",
